feat: refactor time conversion functions into a new module for better organization
